+++ src/client/connections/messaging.rs
@@ -17,7 +17,6 @@ use crate::messaging::{
 use crate::prefix_map::NetworkPrefixMap;
 use crate::types::PublicKey;
 
-use async_recursion::async_recursion;
 use bytes::Bytes;
 use futures::{future::join_all, stream::FuturesUnordered};
 use itertools::Itertools;
@@ -47,6 +46,7 @@ impl Session {
     /// Acquire a session by bootstrapping to a section, maintaining connections to several nodes.
     pub(crate) async fn bootstrap(
         client_pk: PublicKey,
+        genesis_key: bls::PublicKey,
         qp2p_config: QuicP2pConfig,
         err_sender: Sender<CmdError>,
         bootstrap_nodes: BTreeSet<SocketAddr>,
@@ -65,20 +65,15 @@ impl Session {
             .await
             .ok_or(Error::NotBootstrapped)?;
 
-        // *****************************************************
-        // FIXME: receive the network's genesis pk from the user
-        let genesis_pk = bls::SecretKey::random().public_key();
-        // *****************************************************
-
         let session = Session {
             client_pk,
             pending_queries: Arc::new(RwLock::new(HashMap::default())),
             incoming_err_sender: Arc::new(err_sender),
             endpoint,
-            network: Arc::new(NetworkPrefixMap::new(genesis_pk)),
+            network: Arc::new(NetworkPrefixMap::new(genesis_key)),
             aggregator: Arc::new(RwLock::new(SignatureAggregator::new())),
             bootstrap_peer,
-            genesis_pk,
+            genesis_key,
         };
 
         Self::spawn_message_listener_thread(session.clone(), incoming_messages).await;
@@ -88,48 +83,42 @@ impl Session {
 
     /// Utility function that bootstraps a client to a section. If there is a failure then it retries.
     /// After a maximum of three attempts if the boostrap process still fails, then an error is returned.
-    #[async_recursion]
     pub(crate) async fn attempt_bootstrap(
         client_pk: PublicKey,
+        genesis_key: bls::PublicKey,
         qp2p_config: qp2p::Config,
         mut bootstrap_nodes: BTreeSet<SocketAddr>,
         local_addr: SocketAddr,
         err_sender: Sender<CmdError>,
-        attempts: u8,
     ) -> Result<Session, Error> {
-        match Session::bootstrap(
-            client_pk,
-            qp2p_config.clone(),
-            err_sender.clone(),
-            bootstrap_nodes.clone(),
-            local_addr,
-        )
-        .await
-        {
-            Ok(session) => Ok(session),
-            Err(err) => {
-                let attempts = attempts + 1;
-                if let Error::BootstrapToPeerFailed(failed_peer) = err {
-                    // Remove the unresponsive peer we boostrapped to and bootstrap again
-                    let _ = bootstrap_nodes.remove(&failed_peer);
-                }
-                if attempts < NUM_OF_BOOTSTRAPPING_ATTEMPTS {
-                    trace!(
-                        "Error connecting to network! {:?}\nRetrying... ({})",
-                        err,
-                        attempts
-                    );
-                    Self::attempt_bootstrap(
-                        client_pk,
-                        qp2p_config,
-                        bootstrap_nodes,
-                        local_addr,
-                        err_sender,
-                        attempts,
-                    )
-                    .await
-                } else {
-                    Err(err)
+        let mut attempts = 0;
+        loop {
+            match Session::bootstrap(
+                client_pk,
+                genesis_key,
+                qp2p_config.clone(),
+                err_sender.clone(),
+                bootstrap_nodes.clone(),
+                local_addr,
+            )
+            .await
+            {
+                Ok(session) => break Ok(session),
+                Err(err) => {
+                    attempts += 1;
+                    if let Error::BootstrapToPeerFailed(failed_peer) = err {
+                        // Remove the unresponsive peer we boostrapped to and bootstrap again
+                        let _ = bootstrap_nodes.remove(&failed_peer);
+                    }
+                    if attempts < NUM_OF_BOOTSTRAPPING_ATTEMPTS {
+                        trace!(
+                            "Error connecting to network! {:?}\nRetrying... ({})",
+                            err,
+                            attempts
+                        );
+                    } else {
+                        break Err(err);
+                    }
                 }
             }
         }
@@ -161,7 +150,7 @@ impl Session {
             )
         } else {
             // Send message to our bootstrap peer with network's genesis PK.
-            (vec![self.bootstrap_peer], self.genesis_pk)
+            (vec![self.bootstrap_peer], self.genesis_key)
         };
 
         let msg_id = MessageId::new();
@@ -215,7 +204,7 @@ impl Session {
             let mut bootstrapped_peer = BTreeMap::new();
             let _ = bootstrapped_peer.insert(XorName::random(), self.bootstrap_peer);
             // Send message to our bootstrap peer with the network's genesis PK.
-            (bootstrapped_peer, self.genesis_pk)
+            (bootstrapped_peer, self.genesis_key)
         };
 
         // We select the NUM_OF_ELDERS_SUBSET_FOR_QUERIES closest Elders we are querying